fix goroutine leak in client library
Summary: when workflow poller get a workflow task, it handle the workflow task by replay it, during this processe it create a bunch of helper structs one of them is the dispatcher which will then create and attach coroutines to the dispatcher. We need to clean those coroutines (which essentially is goroutine), this diff would fix this leak.

Test Plan: run the sample code, at the end of the code print runtime.NumGoroutine(). Before this fix, the number will increase linearly with the number of workflow/activities executed, with this fix, the goroutine number won't change.

Reviewers: sivakk, maxim, tamer, samar

Reviewed By: samar

Differential Revision: https://code.uberinternal.com/D769230