Client library to use Thrift serializer for Thrift types (#202)
The user inputs/error messages for activity/workflow function defaults to using GOB encoding. We want to use thrift serialization for thrift specific structures, so some of these activities can be cross platform. This is a step towards that, the worker checks if the provided types are thrift types then it would automatically use thrift serialization instead of GOB.