Add an AfterFn mock-delay-returning function (#521)
When combined with a func provided to Return, this can be used to
simulate dynamic / open-ended respons and delay sequences on activity
and workflow functions.

This is particularly useful for running tests with randomized behavior,
as now another important piece of behavior can be varied.  Previously
only the number of calls and the return value could be changed easily.