Decision reordering bug (#529)
* Switched decisionsHelper to LRU semantic

DecisionsHelper keeps an ordered list of DecisionStateMachines. In the cancellation case a new decision is produced by an existing state machine that was added before the cancellation request. It causes the new decision to be generated out of order. Switching the DecisionsHelper to the LRU semantic ensures that any update to a state machine moves it to the end of the ordered list preserving correct order of decisions.