Fix regression in Go SDK where Batch Timer interval for heartbeats is overcalculated (#225)
As part of switching from variables representing time in seconds to time.Duration, we forgot to correctly update the line of code that calculates the interval for sending the next heartbeat. We were multiplying by an extra factor of one billion, which either caused an overflow (so that we were constantly sending heartbeats) or resulted in a large number (where we would never send a heartbeat after the first one).