Remove WorkerWorkflowTasksPerSecond WorkerOption #255
Issue: This worker option throttles the rate at which we process tasks. Unfortunately, this option does not disambiguate between tasks received from the Worker's task queue and tasks received from the Global queue. As a result, a user setting this to a low value can have the unintended consequence of throttling the wrong type of tasks (throttling sticky tasks is bad since those are significantly cheaper to process as all the state needed to process the workflow task is already in memory).

Long-term, we should introduce two separate knobs (one for the Worker task queue and one for the Global task queue). For now, we will just have this not be configurable by the user and go with a default of 100,000.