Fixes issue where a deadlock detected via a Workflow Execution resumed from a Local Activity Completion fails the Workflow (#297)
Fixes regression introduced by Deadlock Detection PR:

A workflow task execution that exceeded 1 second causes a panic.
If the task execution was spawned by the completion of a local activity, that would always fail the Workflow.
This fix alters the code to respect the WorkflowPanicPolicy, even in the case where the panic occurs during Local Activity Result completion. It also adds an integration test to exercise this path.