Replaces time.After() usage with time.NewTimer() to reduce memory pressure due to allocated timers (#300)
See this blog post for an example on when time.After may not be optimal for memory usage purposes: https://medium.com/@oboturov/golang-time-after-is-not-garbage-collected-4cbc94740082

For workflows with a lot of local activities, we will end up unnecessary allocating a whole bunch of timers rather than re-using the same heartbeat one on successive loops. This swaps out time.After() with time.NewTimer() and then explicitly stops the timer when we are done using it.