Error Nil, Not Nil
Various versions of Golang fluctuate by how prevalent of an issue it is, or when dependencies change and forget to use explicit nil.

```
err != nil {

}
```

Is sometimes true when err points to nil in Golang. The only solution is to never use implicit or ambiguous err. Only return err when != nil, and only return keyword `nil` when err was nil. Never unassigned.

It is safe to `return func() error` but only when the underlying function is using explicit returns itself.

```
	if err != nil {
		return err
	}
	return nil
```