Fix an event ordering bug possible with outstanding commands during WF cancel (#382)
Increment the next command id for every command which is cancelled as a result of the workflow itself being cancelled.

This is a bit of a tragic workaround for the event counting Go SDK does which is a fundamentally flawed approach. We
should switch to core-sdk as soon as we can for the state machines.