Ensure default retry policy in test (#462)
* Ensure default retry policy for activity in unit test
This is a breaking change for unit test. Unit tests that assumed no retry might break. For example, if a test setup a mock for activity to return error once. Before this fix, the activity will fail after one attempt. With this fix, the activity will be retried and second attempt will fail because the mock only expect to be called once. Fix to those broken unit test is to either return non retryable error (see NewApplicationError) in those mock, or use a retry policy with MaxAttempts of 1.