enhancement(tracing): export `ActiveSpanContextKey` so consumers can propagate the span context
We trying to trace using opentelemetry bridge https://community.temporal.io/t/trace-temporal-with-opentelemetry-bridge/3226/2. But the each span was in its own trace, after alot of debugging I realized that `ActiveSpanContextKey` is typed `contextKey` so that's why the parent was always `nil`.

Our custom propagater has the `activeSpanContextKey` defined as string where temporal internal have it typed. That's why `createOpenTracingSpan` couldn't get the parent we set in the context.

So this PR suggest exporting it