Address rare custom resolver test race (#725)
* Address rare custom resolver test race

Ensure that the test code doesn't race against grpc connection
establishment in a way that results in test failures. I was able to
reproduce a failure in as few as 12 test runs but sometimes it took more
than 2000 runs to reproduce. That said, this test failed in two of the
last four CI runs so it definitely has an impact.