Adjust HTTP protocol to take Payloads not Payload.
While not all consumers will have the Payloads available those that will can
save on HTTP round trips by sending the whole Payloads object. Temporal Web
will be a good example of where this is possible.

The cost of adding this wrapper for those cases where there is only a single
Payload is so small that the Payloads wrapper is enforced for simplicity.