Returning error from Workflow results in Failure (#838)
The original text, "The error return value is used to indicate an error was encountered during execution and the workflow should be terminated" may be  confusing. 

Although it is true that the Workflow will be terminated in the general sense (i.e., execution stops), the Workflow's status in this case will be failed, not terminated. Since Temporal has a status called "terminated" (i.e., which would result from clicking the Terminate button in the Web UI for a running workflow), I propose changing "terminated" to "failed" here.