Unit test fix for Go 1.19
Go 1.19 (correctly) notices that the following is basically useless as
it tests that an error can be converted to the error interface (which is
always the case).

```
err := NewWorkflowError(...)
target := errors.New("...")

errors.As(err, &target)
```

The broken test was further broken in that it expected the root cause Go
error to make it through the failure conversion machinery intact when in
truth the expectation is that it gets converted to an instance of
ApplicationError.