Workaround for updates delivered in first WFT
If the first WFT includes an update we can have a problem because the
workflow function itself hasn't run yet to register update handlers and
the update will be rejected. Adding this yield allows the scheduler to
execute the workflow function up to the first blocking point, at which
point the handler(s) will be registered by user code.