Send ProtocolMessageCommands where appropriate (#1107)
* Send ProtocolMessageCommands where appropriate

Adds ProtocolMessageCommands to outgoing commands for a WFT response for
update accepted and completed messages but not rejections as rejections
are not written as events. Entries in the message outbox now contain the
message itself as well as a predicate to test history events that may
have been created in response to this message. In this way the messages
and the event expectations are fully defined in internal_update.go and
the event handling and replay code can remain ignorant of the mapping
between message types and the resulting event types.