Add stack trace function for running coroutines
Since Go does not expose a function to take a specific goroutine's stack
trace, we take a full stack trace of all goroutines and search for the
given coroutine via its *coroutineState pointer value and the root
function name. A larger 1MB stackBuf accommodates the full stack trace.
The stack is then formatted similar to getStackTrace(...).