Dump coroutine stack on deadlock timeout (#1356)
* Add identifiable name for root coroutine function

A static function name helps with finding the coroutine's call stack in
a full stack trace of all goroutines. This is useful to print the right
call stack after deadlock detection triggers.

* Add stack trace function for running coroutines

Since Go does not expose a function to take a specific goroutine's stack
trace, we take a full stack trace of all goroutines and search for the
given coroutine via its *coroutineState pointer value and the root
function name. A larger 1MB stackBuf accommodates the full stack trace.
The stack is then formatted similar to getStackTrace(...).

* Set panicError when coroutine deadlocks

The error includes a stack trace of the running coroutine using
getCoroStackTrace, which is much more helpful in debugging than the
stack trace of the deadlock detector shown previously.