Make client Dial context controlable by callers
When clients use `Dial` or `NewClientFromExisting` APIs to start, they
connect to the server and try to load capabilities synchronously. That
being said, they have no control over the timeouts and cancellation of
such requests. They can't extend timeouts to accommodate slow networks
or appropriately cancel those calls.

This commit exposes new `DialContext` and `NewClientFromExistingWithContext`
APIs that take a context as input and propagates it down the stack until
all synchronous server calls so timeout and cancellation can be
controlled by the callers.