testsuite.StartDevServer: Respect timeout during dial (#1498)
testsuite.StartDevServer accepts a context.Context
meant to indicate how long the caller is willing to wait for the result.
This context is used when downloading the Temporal CLI,
but is not considered when tryign to dial to the server in
`waitServerReady`.

Without this, if an attempt to start the server failed
(e.g. because of an invalid `--log-format` argument),
`waitServerReady` will wait for `600 * 100ms = 1 minute`
before returning the failure to the caller.

This changes `waitServerReady` to respect the timeout configured in the
context--checking in on it after each attempt.

This also changes the actual `Dial` attempt to use `DialContext`
with the same context so that if the context expires,
the dial operation also returns early if possible.