New Update-With-Start API (#1731)
Instead of using the ExecuteWorkflow client method, update-with-start is invoked
via a new client method UpdateWithStartWorkflow.

To use this method, first use NewWithStartWorkflowOperation to define
the start-workflow operation. A workflow ID conflict policy is required. Then
call UpdateWithStartWorkflow, passing it an UpdateWithStartWorkflowOptions
containing your WithStartWorkflowOperation, together with an
UpdateWorkflowOptions defining the update operation. This will return an
UpdateHandle. The  WithStartWorkflowOperation exposes a blocking .Get(ctx)
method to obtain the workflow run targeted by the update.