Create v2 Version of datadog tracing interceptor
To completely use v2 of dd-trace-go we must be on the v2 breaking change branch. However, we don't want to cause breaking changes for current users, and we also want to make sure that we respect Go backwards compatibility & semantic versioning.

To that end, the v2 version of this integration uses the most recent version and does not have any remnants of v1. Additionally, it handles the breaking changes between API versions.